+++ src/collective/contact/plonegroup/testing.py
@@ -1,19 +1,26 @@
 # -*- coding: utf-8 -*-
 """Base module for unittesting."""
+import collective.contact.plonegroup
 
-from plone.app.testing import applyProfile
+import pkg_resources
+import unittest2 as unittest
 from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
-from plone.app.testing import login
 from plone.app.testing import PLONE_FIXTURE
 from plone.app.testing import PloneSandboxLayer
-from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID
 from plone.app.testing import TEST_USER_NAME
+from plone.app.testing import applyProfile
+from plone.app.testing import login
+from plone.app.testing import setRoles
 from plone.testing import z2
 
-import collective.contact.plonegroup
-import unittest2 as unittest
+try:
+    pkg_resources.get_distribution('plone.app.contenttypes')
+except pkg_resources.DistributionNotFound:
+    HAS_PA_CONTENTTYPES = False
+else:
+    HAS_PA_CONTENTTYPES = True
 
 
 class CollectiveContactPlonegroupLayer(PloneSandboxLayer):
@@ -30,6 +37,11 @@ class CollectiveContactPlonegroupLayer(PloneSandboxLayer):
     def setUpPloneSite(self, portal):
         """Set up Plone."""
         # Install into Plone site using portal_setup
+
+        # Plone 5 support
+        if HAS_PA_CONTENTTYPES:
+            self.applyProfile(portal, 'plone.app.contenttypes:default')
+
         applyProfile(portal, 'collective.contact.plonegroup:testing')
 
         # Login and create some test content