+++ src/collective/contact/plonegroup/tests/test_subscribers.py
@@ -8,11 +8,11 @@ from collective.contact.plonegroup.interfaces import INotPloneGroupContact
 from collective.contact.plonegroup.interfaces import IPloneGroupContact
 from collective.contact.plonegroup.subscribers import group_deleted
 from collective.contact.plonegroup.testing import IntegrationTestCase
+
+from Products.statusmessages.interfaces import IStatusMessage
 from plone import api
 from plone.app.linkintegrity.exceptions import LinkIntegrityNotificationException
-from plone.app.linkintegrity.interfaces import ILinkIntegrityInfo
 from plone.registry.interfaces import IRegistry
-from Products.statusmessages.interfaces import IStatusMessage
 from zExceptions import Redirect
 from zope.component import getUtility
 
@@ -44,50 +44,6 @@ class TestSubscribers(IntegrationTestCase):
                                   'acontent2',
                                   title='Content 2',
                                   pg_organization=self.contacts[1].UID())
-
-    def test_plonegroupOrganizationRemoved_1(self):
-        """ We cannot remove an organization selected in settings and used in an object """
-        view = self.portal.restrictedTraverse('contacts/%s/department1/delete_confirmation' % PLONEGROUP_ORG)
-        self.assertRaises(LinkIntegrityNotificationException, view.render)
-        storage = ILinkIntegrityInfo(view.REQUEST)
-        breaches = storage.getIntegrityBreaches()
-        self.assertIn(self.contacts[0], breaches)
-        self.assertSetEqual(breaches[self.contacts[0]], set([self.portal['acontent1']]))
-
-    def test_plonegroupOrganizationRemoved_2(self):
-        """ We cannot remove an organization no more selected in settings and used in an object """
-        self.registry[ORGANIZATIONS_REGISTRY] = [self.contacts[0].UID()]  # unselects the contact
-        view = self.portal.restrictedTraverse('contacts/%s/department2/delete_confirmation' % PLONEGROUP_ORG)
-        self.assertRaises(LinkIntegrityNotificationException, view.render)
-        storage = ILinkIntegrityInfo(view.REQUEST)
-        breaches = storage.getIntegrityBreaches()
-        self.assertIn(self.contacts[1], breaches)
-        self.assertSetEqual(breaches[self.contacts[1]], set([self.portal['acontent2']]))
-
-    def test_plonegroupOrganizationRemoved_3(self):
-        """ We can remove an organization no more selected in settings and no more used in an object """
-        self.registry[ORGANIZATIONS_REGISTRY] = [self.contacts[0].UID()]  # unselects the contact
-        self.portal['acontent2'].pg_organization = None
-        self.portal.restrictedTraverse('contacts/%s/department2/delete_confirmation' % PLONEGROUP_ORG)
-
-    def test_plonegroupOrganizationRemoved_4(self):
-        """ We cannot remove an organization selected in settings and used in an object as dict or list """
-        # set uid in dict
-        self.portal['acontent1'].pg_organization = {'uid': self.contacts[0].UID()}
-        view = self.portal.restrictedTraverse('contacts/%s/department1/delete_confirmation' % PLONEGROUP_ORG)
-        self.assertRaises(LinkIntegrityNotificationException, view.render)
-        storage = ILinkIntegrityInfo(view.REQUEST)
-        breaches = storage.getIntegrityBreaches()
-        self.assertIn(self.contacts[0], breaches)
-        self.assertSetEqual(breaches[self.contacts[0]], set([self.portal['acontent1']]))
-        # set uid in list
-        self.portal['acontent2'].pg_organization = [self.contacts[1].UID()]
-        view = self.portal.restrictedTraverse('contacts/%s/department2/delete_confirmation' % PLONEGROUP_ORG)
-        self.assertRaises(LinkIntegrityNotificationException, view.render)
-        storage = ILinkIntegrityInfo(view.REQUEST)
-        breaches = storage.getIntegrityBreaches()
-        self.assertIn(self.contacts[1], breaches)
-        self.assertSetEqual(breaches[self.contacts[1]], set([self.portal['acontent2']]))
 
     def test_plonegroup_contact_transition_1(self):
         """ We cannot deactivate an organization selected in settings """
@@ -133,6 +89,7 @@ class TestSubscribers(IntegrationTestCase):
         class Dummy(object):
             def __init__(self, name):
                 self.principal = name
+
         self.assertIsNone(group_deleted(Dummy('no-underscore')))
         request = self.portal.REQUEST
         smi = IStatusMessage(request)
@@ -140,8 +97,60 @@ class TestSubscribers(IntegrationTestCase):
         self.assertRaises(Redirect, api.group.delete, groupname='%s_director' % uid)
         msgs = smi.show()
         self.assertEqual(msgs[0].message, u"You cannot delete the group '%s_director', linked to used organization "
-                         "'Department 1'." % uid)
+                                          "'Department 1'." % uid)
         api.group.create(groupname='%s_other' % uid)
         api.group.create(groupname='12345_director')
         api.group.delete(groupname='%s_other' % uid)
         api.group.delete(groupname='12345_director')
+
+
+try:
+    from plone.app.linkintegrity.interfaces import ILinkIntegrityInfo
+except ImportError:
+    pass
+else:
+    class TestPlone4Subscribers(TestSubscribers):
+
+        def test_plonegroupOrganizationRemoved_1(self):
+            """ We cannot remove an organization selected in settings and used in an object """
+            view = self.portal.restrictedTraverse('contacts/%s/department1/delete_confirmation' % PLONEGROUP_ORG)
+            self.assertRaises(LinkIntegrityNotificationException, view.render)
+            storage = ILinkIntegrityInfo(view.REQUEST)
+            breaches = storage.getIntegrityBreaches()
+            self.assertIn(self.contacts[0], breaches)
+            self.assertSetEqual(breaches[self.contacts[0]], set([self.portal['acontent1']]))
+
+        def test_plonegroupOrganizationRemoved_2(self):
+            """ We cannot remove an organization no more selected in settings and used in an object """
+            self.registry[ORGANIZATIONS_REGISTRY] = [self.contacts[0].UID()]  # unselects the contact
+            view = self.portal.restrictedTraverse('contacts/%s/department2/delete_confirmation' % PLONEGROUP_ORG)
+            self.assertRaises(LinkIntegrityNotificationException, view.render)
+            storage = ILinkIntegrityInfo(view.REQUEST)
+            breaches = storage.getIntegrityBreaches()
+            self.assertIn(self.contacts[1], breaches)
+            self.assertSetEqual(breaches[self.contacts[1]], set([self.portal['acontent2']]))
+
+        def test_plonegroupOrganizationRemoved_3(self):
+            """ We can remove an organization no more selected in settings and no more used in an object """
+            self.registry[ORGANIZATIONS_REGISTRY] = [self.contacts[0].UID()]  # unselects the contact
+            self.portal['acontent2'].pg_organization = None
+            self.portal.restrictedTraverse('contacts/%s/department2/delete_confirmation' % PLONEGROUP_ORG)
+
+        def test_plonegroupOrganizationRemoved_4(self):
+            """ We cannot remove an organization selected in settings and used in an object as dict or list """
+            # set uid in dict
+            self.portal['acontent1'].pg_organization = {'uid': self.contacts[0].UID()}
+            view = self.portal.restrictedTraverse('contacts/%s/department1/delete_confirmation' % PLONEGROUP_ORG)
+            self.assertRaises(LinkIntegrityNotificationException, view.render)
+            storage = ILinkIntegrityInfo(view.REQUEST)
+            breaches = storage.getIntegrityBreaches()
+            self.assertIn(self.contacts[0], breaches)
+            self.assertSetEqual(breaches[self.contacts[0]], set([self.portal['acontent1']]))
+            # set uid in list
+            self.portal['acontent2'].pg_organization = [self.contacts[1].UID()]
+            view = self.portal.restrictedTraverse('contacts/%s/department2/delete_confirmation' % PLONEGROUP_ORG)
+            self.assertRaises(LinkIntegrityNotificationException, view.render)
+            storage = ILinkIntegrityInfo(view.REQUEST)
+            breaches = storage.getIntegrityBreaches()
+            self.assertIn(self.contacts[1], breaches)
+            self.assertSetEqual(breaches[self.contacts[1]], set([self.portal['acontent2']]))