+++ src/collective/contact/plonegroup/utils.py
@@ -1,16 +1,28 @@
 # -*- coding: utf-8 -*-
+from operator import attrgetter
+from operator import methodcaller
 
 from collective.contact.plonegroup.config import FUNCTIONS_REGISTRY
 from collective.contact.plonegroup.config import ORGANIZATIONS_REGISTRY
 from collective.contact.plonegroup.config import PLONEGROUP_ORG
+
+from Acquisition import aq_get
+from Products.CMFPlone.utils import base_hasattr
+from config import ORGANIZATIONS_REGISTRY
+from config import PLONEGROUP_ORG
 from imio.helpers.content import uuidsToObjects
-from operator import attrgetter
-from operator import methodcaller
 from plone import api
 from plone.app.uuid.utils import uuidToObject
+from plone.behavior.interfaces import IBehavior
+from plone.dexterity.interfaces import IDexterityContent
+from plone.dexterity.interfaces import IDexterityFTI
 from zope.annotation.interfaces import IAnnotations
 from zope.component import getUtility
 from zope.globalrequest import getRequest
+from zope.schema import getFieldsInOrder
+from zope.schema.interfaces import IChoice
+from zope.schema.interfaces import ICollection
+from zope.schema.interfaces import IText
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
@@ -221,3 +233,83 @@ def select_organization(org_uid, remove=False):
     else:
         plonegroup_organizations.append(org_uid)
     api.portal.set_registry_record(ORGANIZATIONS_REGISTRY, plonegroup_organizations)
+
+
+def search_value_in_objects(s_obj, ref, p_types=[], type_fields={}):
+    """
+        Searching a value (reference to an object like id or uid) in fields of objects.
+        Parameters:
+            * s_obj : the object that is maybe referenced in another objects fields
+            * ref : the value to search in field
+            * p_types : portal_types that will be only searched
+            * type_fields : dict containing as key portal_type and as value a list of fields that must be searched.
+                            If a portal_type is not given, all fields will be searched
+    """
+    # we check all dexterity objects fields to see if ref is used in
+    # we can't check only fields using plonegroup vocabulary because maybe another vocabulary name is used
+    # this can be long but this operation is not made so often
+
+    request = aq_get(s_obj, 'REQUEST', None)
+    if not request:
+        return
+    try:
+        catalog = api.portal.get_tool('portal_catalog')
+    except api.portal.CannotGetPortalError:
+        # When deleting site, the portal is no more found...
+        return
+
+    def list_fields(ptype, filter_interfaces=(IText, ICollection, IChoice)):
+        """ return for the portal_type the selected fields """
+        if ptype not in type_fields:
+            type_fields[ptype] = []
+            fti = getUtility(IDexterityFTI, name=ptype)
+            for name, fld in getFieldsInOrder(fti.lookupSchema()):
+                for iface in filter_interfaces:
+                    if iface.providedBy(fld):
+                        type_fields[ptype].append(name)
+                        break
+            # also lookup behaviors
+            for behavior_id in fti.behaviors:
+                behavior = getUtility(IBehavior, behavior_id).interface
+                for name, fld in getFieldsInOrder(behavior):
+                    for iface in filter_interfaces:
+                        if iface.providedBy(fld):
+                            type_fields[ptype].append(name)
+                            break
+        return type_fields[ptype]
+
+    def check_value(val):
+        if isinstance(val, basestring) and val == ref:
+            return True
+        return False
+
+    def check_attribute(val):
+        """ check the attribute value and walk in it """
+        if isinstance(val, dict):
+            for v in val.values():
+                res = check_attribute(v)
+                if res:
+                    return res
+        elif base_hasattr(val, '__iter__'):
+            for v in val:
+                res = check_attribute(v)
+                if res:
+                    return res
+        elif check_value(val):
+            res = [val]
+            return res
+        return []
+
+    results = set()
+    for brain in catalog.unrestrictedSearchResults(portal_types=p_types,
+                                                   object_provides=IDexterityContent.__identifier__):
+        obj = brain._unrestrictedGetObject()
+        ptype = obj.portal_type
+        for attr in list_fields(ptype):
+            if base_hasattr(obj, attr):
+                res = check_attribute(getattr(obj, attr))
+                if res:
+                    results.add(obj)
+                    break
+
+    return list(results)